Added location (needed it in validation). Updated datacontext, added getinventory in repo, interface and controller.

+++ AdventureWork/Models/ProductInventory.cs
@@ -6,10 +6,10 @@ namespace AdventureWork.Models
     public class ProductInventory
     {
         public int ProductID { get; set; }
-        public int LocationID { get; set; }
+        public Int16 LocationID { get; set; }
         public string Shelf { get; set; }
-        public int Bin { get; set; }
-        public int Quantity { get; set; }
+        public Byte Bin { get; set; }
+        public Int16 Quantity { get; set; }
         public Guid rowguid { get; set; }
         public DateTime ModifiedDate { get; set; }
